pyproject.toml
[tool.poetry]
name = "centaur_workspace"
version = "0.1.0"
description = "CrewAI-based workspace for AI agent development and interaction"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "centaur_workspace" }
]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = "^0.36.1"
crewai_tools = "^0.4.8"
openai = "^1.35.14"
pyyaml = "^6.0"
python-dotenv = "^1.0.1"
pysbd = "^0.3.4"
google-auth = "^2.32.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.137.0"
pydantic = "^2.0.0"
posthog = ">=3.0.2,<4.0.0"
anthropic = "^0.31.2"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"
black = "^23.0.0"
pre-commit = "^2.21.0"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.8"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "manual: marks tests as manual tests",
]
filterwarnings = [
    "ignore:.*_upb._message.*:DeprecationWarning",
    "ignore:.*datetime.utcnow.*:DeprecationWarning",
    "ignore:.*Pydantic V1 style.*:DeprecationWarning",
    "ignore::DeprecationWarning",
]

[tool.poetry.scripts]
centaur = "centaur_workspace.main:main"
